requirements.txt
# Core dependencies for the news sentiment analysis system
streamlit>=1.28.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# News fetching and API integration
requests>=2.31.0
python-dotenv>=1.0.0

# Data visualization
plotly>=5.15.0

# Machine learning and text processing
nltk>=3.8.0
scipy>=1.11.0

# Optional: Uncomment for Spark-based version
# pyspark>=3.4.0
# findspark>=2.0.0

# Development tools (optional)
# pytest>=7.4.0
# black>=23.7.0
